+++ sibl_gui/components/addons/onlineUpdater/downloadManager.py
@@ -520,7 +520,7 @@ class DownloadManager(foundations.ui.common.QWidgetFactory(uiFile=UI_FILE)):
 		self.__currentRequest = self.__networkAccessManager.get(QNetworkRequest(QUrl(self.__requests.pop())))
 
 		self.__currentFilePath = os.path.join(self.__downloadDirectory,
-											os.path.basename(foundations.strings.toUnicode(self.__currentRequest.url().path())))
+											os.path.basename(foundations.strings.toString(self.__currentRequest.url().path())))
 		if foundations.common.pathExists(self.__currentFilePath):
 			LOGGER.info("{0} | Removing '{1}' local file from previous online update!".format(
 			self.__class__.__name__, os.path.basename(self.__currentFilePath)))
@@ -552,7 +552,7 @@ class DownloadManager(foundations.ui.common.QWidgetFactory(uiFile=UI_FILE)):
 																									bytesTotal))
 
 		self.Current_File_label.setText("Downloading: '{0}'.".format(
-		os.path.basename(foundations.strings.toUnicode(self.__currentRequest.url().path()))))
+		os.path.basename(foundations.strings.toString(self.__currentRequest.url().path()))))
 		self.Download_progressBar.setRange(0, bytesTotal)
 		self.Download_progressBar.setValue(bytesReceived)